Split policy.py into model and parser modules.
Splitting the code into model and parser has the following benefits:

- clarifies the code
- allows for different data sources to load the policies (e.g., YAML, database tables)
- separates the model out so that unit tests can be added for the model only
- allows for model-level changes for different use cases that would otherwise impact the Capirca grammar
